deps/v8/test/inspector/debugger/get-properties-paused-expected.txt

Checks Runtime.getProperties method while debugger is paused.

Running test: testObject5
  foo own string cat
Internal properties
  [[PrimitiveValue]] number 5
  [[Prototype]] object undefined

Running test: testNotOwn
  __defineGetter__ inherited function undefined
  __defineSetter__ inherited function undefined
  __lookupGetter__ inherited function undefined
  __lookupSetter__ inherited function undefined
  __proto__ inherited no value, getter, setter
  a own number 2
  b own no value, getter, setter
  c inherited number 4
  constructor inherited function undefined
  d inherited no value, getter
  hasOwnProperty inherited function undefined
  isPrototypeOf inherited function undefined
  propertyIsEnumerable inherited function undefined
  toLocaleString inherited function undefined
  toString inherited function undefined
  valueOf inherited function undefined
Internal properties
  [[Prototype]] object undefined

Running test: testAccessorsOnly
  b own no value, getter, setter
  d own no value, setter

Running test: testArray
  0 own string red
  1 own string green
  2 own string blue
  length own number 3
Internal properties
  [[Prototype]] object undefined

Running test: testBound
  length own number 0
  name own string bound Number
Internal properties
  [[BoundArgs]] object undefined
  [[BoundThis]] object undefined
  [[Prototype]] function undefined
  [[TargetFunction]] function undefined

Running test: testObjectThrowsLength
  length own no value, getter
Internal properties
  [[Prototype]] object undefined

Running test: testTypedArrayWithoutLength
Internal properties
  [[Prototype]] object undefined

Running test: testArrayBuffer

Running test: testArrayBufferWithBrokenUintCtor
Internal properties
  [[ArrayBufferByteLength]] number 7
  [[ArrayBufferData]] string 0x...
  [[Int8Array]] object undefined
  [[Prototype]] object undefined
  [[Uint8Array]] object undefined
